Fixing a couple of warnings
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@7170 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/heartbeat/win_perf_counters.cpp
@@ -58,7 +58,7 @@ void install_windows_counters()
 void update_windows_counters(boost::uint64_t value)
 {
     // Set raw counter data for queue length.
-    ULONG status = PerfSetULongCounterValue(HPXHeartBeat, queue_counter, 1, value);
+    ULONG status = PerfSetULongCounterValue(HPXHeartBeat, queue_counter, 1, ULONG(value));
     if (status != ERROR_SUCCESS) {
         std::cerr << "PerfSetCounterRefValue for 'sum_queue_counter' failed "
                      "with error code: "
@@ -67,7 +67,7 @@ void update_windows_counters(boost::uint64_t value)
     }
 
     // Set raw counter data for average queue length.
-    status = PerfSetULongCounterValue(HPXHeartBeat, avg_queue_counter, 2, value);
+    status = PerfSetULongCounterValue(HPXHeartBeat, avg_queue_counter, 2, ULONG(value));
     if (status != ERROR_SUCCESS) {
         std::cerr << "PerfSetCounterRefValue for 'avg_queue_counter' failed "
                      "with error code: "
